03_mysql_registration_bot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.1.31
mysql-connector-python==9.2.0
python-telegram-bot==13.15
pytz==2025.1
six==1.17.0
tornado==6.1
tzlocal==5.3.1
